src/pyk/{cli_utils => cli.utils}: move to subdirectory
