+++ src/pyk/kbuild/utils.py
@@ -5,7 +5,7 @@ import shutil
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, final
 
-from pyk.cli_utils import check_dir_path, check_file_path, run_process
+from ..cli.utils import check_dir_path, check_file_path, run_process
 
 if TYPE_CHECKING:
     from collections.abc import Iterable
