+++ src/pyk/kore/tools.py
@@ -4,7 +4,7 @@ from enum import Enum
 from pathlib import Path
 from tempfile import NamedTemporaryFile
 
-from ..cli_utils import check_dir_path, run_process
+from ..cli.utils import check_dir_path, run_process
 from .syntax import Pattern
