+++ src/pyk/ktool/kompile.py
@@ -13,7 +13,7 @@ from pathlib import Path
 from subprocess import CalledProcessError
 from typing import TYPE_CHECKING, final
 
-from ..cli_utils import abs_or_rel_to, check_dir_path, check_file_path, run_process
+from ..cli.utils import abs_or_rel_to, check_dir_path, check_file_path, run_process
 
 if TYPE_CHECKING:
     from collections.abc import Iterable, Mapping
